More small fixed to Levermore1D
Former-commit-id: 8cbce025256f96ed2a403145dae4afcb2e41b052

+++ src_2D/Levermore1D/UnitTests/test_Levermore1D_pState.cc
@@ -200,7 +200,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       W1[i] = a;
       W2[i] = b;
     }
@@ -209,7 +209,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       ensure_distance("W3 = W1 + W2", W3[i], a+b, fabs(a+b)*tol);
     }
   }
@@ -227,7 +227,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       W1[i] = a;
       W2[i] = b;
     }
@@ -236,7 +236,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
 
       ensure_distance("W2 += W1",W2[i],a+b,fabs(a+b)*tol);
       ensure_distance("W3 = (W2+=W1)",W3[i],a+b,fabs(a+b)*tol);
@@ -256,7 +256,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       W1[i] = a;
       W2[i] = b;
     }
@@ -265,7 +265,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       ensure_distance("W3 = W1 - W2", W3[i], a-b, fabs(a-b)*tol);
     }
   }
@@ -283,7 +283,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       W1[i] = a;
       W2[i] = b;
     }
@@ -292,7 +292,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
 
       ensure_distance("W2 -= W1",W2[i],b-a,fabs(b-a)*tol);
       ensure_distance("W3 = (W2-=W1)",W3[i],b-a,fabs(b-a)*tol);
@@ -312,7 +312,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       dot += (a*b);
       W1[i] = a;
       W2[i] = b;
@@ -338,7 +338,7 @@ namespace tut
 
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       W1[i] = a;
       W2[i] = b;
     }
@@ -346,14 +346,14 @@ namespace tut
     W3 = W1 ^ W2;
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       ensure_distance("W3 = W1^W2)", W3[i], a*b, fabs(a*b)*tol);
     }
 
     W3 = W2 ^ W1;
     for(i=1;i<=Levermore1D_pState::get_length();++i) {
       a = pow((double)i,1.23456) / 98.765;
-      b = sqrt((double)i) * exp (i);
+      b = sqrt((double)i) * exp((double)i);
       ensure_distance("W3 = W1^W2)", W3[i], a*b, fabs(a*b)*tol);
     }
   }
